python/examples/llama-index/backend/requirements.txt
aiohappyeyeballs==2.3.6
aiohttp==3.10.3
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
-e git+https://github.com/Arize-ai/openinference.git@76d7db642903a14d58b0872ea8d22f360f9d83a4#egg=app&subdirectory=python\examples\llama-index\backend
arize==7.20.1
arrow==1.3.0
attrs==24.2.0
beautifulsoup4==4.12.3
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
coloredlogs==15.0.1
dataclasses-json==0.6.7
Deprecated==1.2.14
diff-match-patch==20230430
dirtyjson==1.0.8
distro==1.9.0
docx2txt==0.8
Faker==25.9.2
fastapi==0.109.2
filelock==3.15.4
fqdn==1.5.1
frozenlist==1.4.1
fsspec==2024.6.1
googleapis-common-protos==1.63.2
greenlet==3.0.3
griffe==0.36.9
grpcio==1.65.5
guardrails-ai==0.5.4
guardrails-api-client==0.3.9
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.24.5
humanfriendly==10.0
idna==3.7
importlib_metadata==8.0.0
isoduration==20.11.0
Jinja2==3.1.4
jiter==0.5.0
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonref==1.1.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
langchain-core==0.2.33
langsmith==0.1.99
litellm==1.43.16
llama-index==0.10.44
llama-index-agent-openai==0.2.7
llama-index-cli==0.1.13
llama-index-core==0.10.44
llama-index-embeddings-openai==0.1.11
llama-index-indices-managed-llama-cloud==0.1.6
llama-index-legacy==0.9.48.post2
llama-index-llms-openai==0.1.26
llama-index-multi-modal-llms-openai==0.1.9
llama-index-program-openai==0.1.6
llama-index-question-gen-openai==0.1.3
llama-index-readers-file==0.1.33
llama-index-readers-llama-parse==0.1.6
llama-parse==0.4.9
llamaindex-py-client==0.1.19
lxml==4.9.4
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.3
mdurl==0.1.2
multidict==6.0.5
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.3
nltk==3.8.1
numpy==2.0.1
openai==1.41.0
openinference-instrumentation==0.1.15
openinference-instrumentation-guardrails==0.1.1
openinference-instrumentation-llama-index==2.2.4
openinference-semantic-conventions==0.1.10
opentelemetry-api==1.26.0
opentelemetry-exporter-otlp==1.26.0
opentelemetry-exporter-otlp-proto-common==1.26.0
opentelemetry-exporter-otlp-proto-grpc==1.26.0
opentelemetry-exporter-otlp-proto-http==1.26.0
opentelemetry-instrumentation==0.47b0
opentelemetry-proto==1.26.0
opentelemetry-sdk==1.26.0
opentelemetry-semantic-conventions==0.47b0
orjson==3.10.7
packaging==24.1
pandas==2.2.2
pillow==10.4.0
protobuf==4.25.4
pyarrow==17.0.0
pydantic==2.8.2
pydantic_core==2.20.1
pydash==7.0.7
Pygments==2.18.0
PyJWT==2.9.0
pypdf==4.3.1
pyreadline3==3.4.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.2
referencing==0.35.1
regex==2023.12.25
requests==2.32.3
requests-futures==1.0.0
rfc3339-validator==0.1.4
rfc3987==1.3.8
rich==13.7.1
rpds-py==0.20.0
rstr==3.2.2
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.32
starlette==0.36.3
striprtf==0.0.26
tenacity==8.5.0
tiktoken==0.7.0
tokenizers==0.20.0
tqdm==4.66.5
typer==0.12.4
types-python-dateutil==2.9.0.20240316
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.1
uri-template==1.3.0
urllib3==2.2.2
uvicorn==0.23.2
watchfiles==0.23.0
webcolors==24.8.0
websockets==12.0
wrapt==1.16.0
yarl==1.9.4
zipp==3.20.0
